+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `test_tug_o_war.py` smoketests of basic LQ example
 - `test_unicycle.py` smoketest of basic ILQ example
 - `test_constrained_integrators.py` smoketest of basic AL example
+- `frontend/solvers.py` a `SolveResult.__str__` method to unify result printing 
 - `scripts/visuals/lady_bandit_guard_plot.py` for generating optional documentation plots outside the package
 - `scripts/visuals/lady_bandit_guard_nonlinear_plot.py` for generating optional plots for the nonlinear LBG example
 - `matplotlib` to optional dependencies to generate plots from `scripts/visuals`